Fixing compilation with -Wconversion on 64-bit systems.

+++ examples/containers/deque2.cpp
@@ -34,7 +34,7 @@ int main()
     stxxl::deque_iterator<deque> deque_iterator = my_deque.begin();
 
     // Access random element x at position p(x) in the deque
-    p = rand32() % number_of_elements;
+    p = (unsigned int)(rand32() % number_of_elements);
     x = my_deque[p];
 
     // Count number of smaller elements from the front to p(x) - 1
